Split tests of TasksAPI into several test cases
