+++ tasks/tests/tests_anonymous_user_test_tasks_api.py
@@ -1,0 +1,38 @@
+from rest_framework.test import APITestCase
+from rest_framework import status
+from tasks.models import Task
+from projects.models import Project
+
+
+class AnonymousTasksTest(APITestCase):
+    fixtures = ['users.json', 'projects.json', 'tasks.json']
+    base_url = "/api/v1/tasks/"
+
+    def setUp(self):
+        self.project = Project.objects.first()
+        self.task = self.project.task_set.first()
+        self.task_url = self.base_url + '{0}/'.format(self.task.id)
+        self.tasks = self.project.task_set.all()
+
+    def test_tasks_list(self):
+        url_params = dict(project_id=self.project.id)
+        r = self.client.get(self.base_url, data=url_params)
+        self.assertEqual(r.status_code, status.HTTP_403_FORBIDDEN)
+
+    def test_get_task(self):
+        r = self.client.get(self.task_url)
+        self.assertEqual(r.status_code, status.HTTP_403_FORBIDDEN)
+
+    def test_add_new_task(self):
+        data = dict(title="my new task", project=self.project.id, header=self.project.header.id)
+        r = self.client.post(self.base_url, data)
+        self.assertEqual(r.status_code, status.HTTP_403_FORBIDDEN)
+
+    def test_patch_task(self):
+        data = dict(title="my updated task")
+        r = self.client.patch(self.task_url, data)
+        self.assertEqual(r.status_code, status.HTTP_403_FORBIDDEN)
+
+    def test_delete_task(self):
+        r = self.client.delete(self.task_url)
+        self.assertEqual(r.status_code, status.HTTP_403_FORBIDDEN)
