+++ tasks/tests/tests_project_header_test_tasks_api.py
@@ -1,0 +1,28 @@
+from rest_framework.test import APITestCase
+from rest_framework import status
+from tasks.models import Task
+from projects.models import Project
+
+
+class ProjectHeaderTasksTest(APITestCase):
+    fixtures = ['users.json', 'projects.json', 'tasks.json']
+    base_url = "/api/v1/tasks/"
+
+    def setUp(self):
+        self.project = Project.objects.first()
+        self.task = self.project.task_set.first()
+        self.task_url = self.base_url + '{0}/'.format(self.task.id)
+        self.tasks = self.project.task_set.all()
+        self.client.login(username=self.project.header.username, password="password")
+
+    def test_get_list_tasks(self):
+        url_params = dict(project_id=self.project.id)
+        r = self.client.get(self.base_url, data=url_params)
+        self.assertEqual(r.status_code, status.HTTP_200_OK)
+        self.assertSetEqual(set(self.tasks.values_list('id', flat=True)), set(map(lambda x: x['id'], r.json())))
+
+    def test_add_new_task(self):
+        data = dict(title="my new task", project=self.project.id, header=self.project.header.id)
+        r = self.client.post(self.base_url, data)
+        self.assertEqual(r.status_code, status.HTTP_201_CREATED)
+        self.assertEqual(Task.objects.filter(id=r.json()['id']).count(), 1)
